% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zman_trajectory_analysis.R
\name{calculate_corr_genes}
\alias{calculate_corr_genes}
\title{Calculate correlated genes along time}
\usage{
calculate_corr_genes(new_id, smoothed_res, method = "spearman")
}
\arguments{
\item{new_id}{ID from metacell object}

\item{smoothed_res}{Result object from smooth_zman_trajectory()}

\item{method}{Defaults to "spearman", but can be "pearson" or "kendall" as seen from cor.test function}
}
\description{
This function calculates the correlation between the gene expression and the time
}
\section{Trajectory Analysis}{
NA
}

\examples{
# Smooth AUC trajectory:
# First we filter genes by dispersion to reduce the number of genes to more significant ones:
select_GO_mc <- get_GO_exp(select_exprs,gene_type = "gene_names", organism = "mouse", takelog=F)
# The threshold is set roughly at 90\% quantile of the mc, since metacells have fewer cells and pooled expression, we expect in general lower dispersion values
filtered_GO_mc <- filter_exp(select_GO_mc, dispersion_threshold=0.05, threadnum = 1)
NK_smoothed_res = smooth_zman_trajectory(filtered_GO_mc, NK_mc_exprs, ref_k = 4)
NK_corr = calculate_corr_genes(new_id, NK_smoothed_res, "spearman") # list containing both correlation and pvalues vectors for each gene
}
