% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zman_trajectory_analysis.R
\name{compute_mc_cdf}
\alias{compute_mc_cdf}
\title{Computing the CDF / AUC value for each metacell}
\usage{
compute_mc_cdf(
  new_id,
  time_per_cell,
  select_mcs = 1:37,
  mc_annotations = NULL,
  time_points = c("12H", "24H", "36H"),
  time_for_auc = c(0, 12, 24, 36),
  min_cell_per_mc_mouse = 10,
  Gate = NULL
)
}
\arguments{
\item{new_id}{id for loading the relevant metacell objects for scdb_mc and scdb_mat}

\item{time_per_cell}{This is the annotated dataframe for each single cell with the time group assignment}

\item{select_mcs}{Take the relevant mc id for normalization. Here we take the T cells and NK cells together for normalization as an example}

\item{mc_annotations}{Dataframe with cell type annotations for each metacell id}

\item{time_points}{A vector containing the unique time points that you have in the experiment (ie c("12H", "24H", "36H"))}

\item{time_for_auc}{Corresponding numeric vector that provides the time values with an additional start of 0 for calculation of AUC (ie c(0,12,24,36))}

\item{min_cell_per_mc_mouse}{Threshold for minimum number of cells per mc of a mouse. Set to 10 for the NK example dataset.}

\item{Gate}{Use CD45 gating for normalization. Defaults to NULL if all cells have been gated as CD45. The column "Gating" is used for subseting}
}
\description{
This function calculates the AUC time value for each metacell using the time bin assignment information from FACS
}
\section{Trajectory Analysis}{
NA
}

\examples{
# Assuming `my_data` is your FACS data
new_id = "T_clean"
time_per_cell = well_fcs_GBM_T_time
mc_cdf <- compute_mc_cdf(new_id, time_per_cell, select_mcs = 1:37, mc_annotations = mc_annotations, time_points = c("12H","24H","36H"), time_for_auc = c(0,12,24,36), Gate="CD45_high")
}
