% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_gene_expression.R
\name{ensembl_to_mgi}
\alias{ensembl_to_mgi}
\title{Gene name conversion: ENSEMBL -> MGI}
\usage{
ensembl_to_mgi(ensembl_names, gene_table)
}
\arguments{
\item{ensembl_names}{The list of gene names in ENSEMBL. (Mouse)}

\item{gene_table}{mm_IDs (loaded from preprocessed table)}
}
\description{
Provides gene name conversion for ease of processing.
}
\section{Preprocessing}{

}

\examples{
ensembl_to_mgi(ensembl_names, gene_table = mm_IDs)
}
