% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_gene_expression.R
\name{entrez_to_mgi}
\alias{entrez_to_mgi}
\title{Gene name conversion: ENTREZ -> MGI}
\usage{
entrez_to_mgi(entrez_names, gene_table)
}
\arguments{
\item{entrez_names}{The list of gene names in ENTREZ. (Mouse)}

\item{gene_table}{mm_IDs (loaded from preprocessed table)}
}
\description{
Provides gene name conversion for ease of processing.
}
\section{Preprocessing}{
NA
}

\examples{
entrez_to_mgi(entrez_list, gene_table = mm_IDs)
}
