% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_assignment.R
\name{FACS_model}
\alias{FACS_model}
\title{FACS_model: Classify cells based on flow cytometry data}
\usage{
FACS_model(
  well_fcs_mc,
  fluorophores = c("log_BV711.A", "log_BUV737.A", "log_BB515.A", "log_PE.A"),
  sd_threshold = c(3, 3, 3, 3),
  timebins = c("12H", "24H", "36H", "48H")
)
}
\arguments{
\item{well_fcs_mc}{Data frame containing flow cytometry data}

\item{fluorophores}{Vector of fluorophore channels to be considered. Defaults to c('log_BV711-A','log_BUV737-A','log_BB515-A','log_PE-A')}

\item{sd_threshold}{Vector of standard deviation thresholds corresponding to the specified channels. Defaults to c(3, 3, 3, 3)}

\item{timebins}{Vector of the corresponding timebins for each fluorophore}
}
\value{
A data frame of cell metadata with assignment columns (1 or -1 for each fluorophore/timebin) and a final timebin assignment column named 'group'
}
\description{
FACS_model: Classify cells based on flow cytometry data
}
\section{Timebin Assignment}{
NA
}

\examples{
# Assuming 'my_data' is your flow cytometry data frame
result <- FACS_model(my_data)
result <- FACS_model(my_data, fluorophores = c('log_BV711-A','log_BUV737-A','log_BB515-A','log_PE-A'), sd_threshold =  c(3,3,3,3),
                     timepoints = c("12H","24H","36H","48H"))
}
