% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_assignment.R
\name{FACS_model_eval}
\alias{FACS_model_eval}
\title{Evaluate FACS Models for Various Fluorophores}
\usage{
FACS_model_eval(
  well_fcs_mc,
  fluorophores = c("log_BV711.A", "log_BUV737.A", "log_BB515.A", "log_PE.A")
)
}
\arguments{
\item{well_fcs_mc}{Data frame containing the FACS data with columns for different fluorophores.}
}
\value{
Visualizations for assessing the normality of the fitted model
}
\description{
This function evaluates models for different fluorophores including BV711, PE, BB515, and BUV737
using Generalized Linear Models (GLM).
At the end, it returns the filtered dataframe and plots for the GLM fits.
}
\section{Timebin Assignment}{
NA
}

\examples{
# Assuming `my_data` is your FACS data
FACS_model_eval(my_data)
}
