% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_gene_expression.R
\name{filter_exp}
\alias{filter_exp}
\title{Filter genes}
\usage{
filter_exp(cell_gene_matrix, dispersion_threshold = 10, threads = 2)
}
\arguments{
\item{cell_gene_matrix}{A cell (row) by gene (col) matrix. Typically the result returned from get_GO_exp.}

\item{dispersion_threshold}{Defaults to 10. Depending on the type of normalization / data, the ratio should be changed to retain a few hundred genes.}

\item{threads}{The number of threads to use.}
}
\description{
This function reduces the dimensionality by filtering out low dispersion ratio genes.
}
\section{Preprocessing}{

}

\examples{
subset_GO_data <- get_GO_exp(data, gene_type = "gene_names", organism = "human", takelog = F)
filtered_GO_data <- filter_exp(subset_GO_data, dispersion_threshold = 10, threads = 2)

}
\keyword{filter_exp}
