% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zman_trajectory_analysis.R
\name{normalize_mc_exprs}
\alias{normalize_mc_exprs}
\title{Re-normalizing MC data for Zman trajectory}
\usage{
normalize_mc_exprs(
  new_id,
  mc_annotations,
  mc2d_auc_time,
  select_celltypes = c("chemotactic", "cytotoxic", "intermediate", "dysfunctional"),
  norm_by_mc_meansize = TRUE
)
}
\arguments{
\item{new_id}{id for loading the relevant metacell objects for scdb_mc and scdb_mat}

\item{mc_annotations}{Dataframe with cell type annotations for each metacell id}

\item{mc2d_auc_time}{Dataframe containing the coordinates of the metacells and its corresponding AUC values and cell type information}

\item{select_celltypes}{Select cell types used for constructing the trajectory. i.e. we subset NKs or myeloid cells separately to look at their temporal profile.}
}
\description{
This function prepares the data required for smoothing the Zman AUC derived trajectory
}
\section{Trajectory Analysis}{
NA
}

\examples{
# Normalize mc expression data for selected lineage of cell types
NK_mc_exprs = normalize_mc_exprs(new_id, mc_annotations, mc_cdf$mc2d_auc_time, select_celltypes = c("chemotactic", "cytotoxic", "intermediate", "dysfunctional"))
}
