% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_assignment.R
\name{plot_FACS_eval}
\alias{plot_FACS_eval}
\title{plot_FACS_eval: Plot FACS data after assignment and check if assigned variable are good}
\usage{
plot_FACS_eval(
  well_fcs_mc,
  fluorophores = c("log_BV711.A", "log_BUV737.A", "log_BB515.A", "log_PE.A"),
  timebins = c("12H", "24H", "36H", "48H"),
  y_axis = "log_APC.A"
)
}
\arguments{
\item{well_fcs_mc}{Data frame containing FACS data}

\item{fluorophores}{Vector of channels to be plotted}

\item{timebins}{Vector of the corresponding timebins for each fluorophore}

\item{y_axis}{String of fluorophore to be plotted against in the y-axis. Defaults to 'log_APC-A'}
}
\value{
A ggplot object displaying the data
}
\description{
plot_FACS_eval: Plot FACS data after assignment and check if assigned variable are good
}
\examples{
# Assuming 'data' is your FACS data and 'assignment_data' contains the assignments
plot_FACS_eval_3(data, assignment_data)
plot_FACS_eval_3(data, assignment_data, channels = c('custom_channel1', 'custom_channel2'))
}
