% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_assignment.R
\name{plot_FACS_histogram}
\alias{plot_FACS_histogram}
\title{Plot Histograms for FACS Data}
\usage{
plot_FACS_histogram(
  well_fcs_mc,
  fluorophores = c("log_BV711-A", "log_BUV737-A", "log_BB515-A", "log_PE-A")
)
}
\arguments{
\item{well_fcs_mc}{Data frame containing the FACS data with columns for different fluorophores.}
}
\description{
This function plots histograms for various fluorophores including BV711, PE, BB515, and BUV737.
It uses ggplot2 for the plotting.
}
\section{Timebin Assignment}{
NA
}

\examples{
# Assuming `my_data` is your FACS data
plot_FACS_histogram(my_data)
}
