% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zman_trajectory_analysis.R
\name{plot_smoothed_trajectory}
\alias{plot_smoothed_trajectory}
\title{Plot gene expression across time for insights and visualization}
\usage{
plot_smoothed_trajectory(smoothed_res, mc_cdf_res, mc_only = FALSE)
}
\arguments{
\item{smoothed_res}{Result object from smooth_zman_trajectory()}

\item{mc_only}{Defaults to FALSE, showing the trajectory arrows. If TRUE, it only shows the graph of metacells}

\item{mc_cdf}{Result object from compute_mc_cdf()}
}
\description{
This function utilizes the smoothed trajectory to plot the gene expression trend across time using loess fitting
}
\section{Trajectory Analysis}{
NA
}

\examples{
filtered_GO_mc <- filter_exp(select_GO_mc, dispersion_threshold=0.05, threadnum = 1)
NK_smoothed_res = smooth_zman_trajectory(filtered_GO_mc, NK_mc_exprs$select_df, ref_k = 4)
NK_predicted_res = predict_expression_along_time(new_id, NK_smoothed_res, out_len = 36, loess_degree=1, loess_span = .9)
plot_zman_genes_heatmap(NK_predicted_res, NK_smoothed_res, c("Tigit","Xcl1","Pmepa1","Igflr1", "Gzmc"), c("Ccl3","Prf1","Gzma", "Gzmb","Nkg7"), k = 2)
plot_smoothed_trajectory(NK_smoothed_res, mc_cdf, mc_only=FALSE)
}
