% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zman_trajectory_analysis.R
\name{plot_zman_genes_heatmap}
\alias{plot_zman_genes_heatmap}
\title{Plot gene expression across time for insights and visualization}
\usage{
plot_zman_genes_heatmap(
  predicted_exprs,
  smoothed_res,
  up_regulated_genes,
  down_regulated_genes,
  k = 2,
  row_size = 5.5
)
}
\arguments{
\item{predicted_exprs}{Result object from predict_expression_along_time()}

\item{up_regulated_genes}{Genes identified from correlation analysis that are up-regulated along time}

\item{down_regulated_genes}{Genes identified from correlation analysis that are up-regulated along time}

\item{k}{Number of gene groups for clustering separation, defaults to 2 for up- and down- regulated genes}

\item{row_size}{Size of text for genes (Defaults to 5.5)}
}
\description{
This function utilizes the smoothed trajectory to plot the gene expression trend across time using loess fitting
}
\section{Trajectory Analysis}{
NA
}

\examples{
filtered_GO_mc <- filter_exp(select_GO_mc, dispersion_threshold=0.05, threadnum = 1)
NK_smoothed_res = smooth_zman_trajectory(filtered_GO_mc, NK_mc_exprs$select_df, ref_k = 4)
NK_predicted_res = predict_expression_along_time(new_id, NK_smoothed_res, out_len = 36, loess_degree=1, loess_span = .9)
plot_zman_genes_heatmap(NK_predicted_res, NK_smoothed_res, c("Tigit","Xcl1","Pmepa1","Igflr1", "Gzmc"), c("Ccl3","Prf1","Gzma", "Gzmb","Nkg7"), k = 2)
}
