% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zman_trajectory_analysis.R
\name{predict_expression_along_time}
\alias{predict_expression_along_time}
\title{Predict gene expression across time for insights and visualization}
\usage{
predict_expression_along_time(
  new_id,
  smoothed_res,
  out_len = 36,
  loess_degree = 1,
  loess_span = 0.9,
  reverse_traj = FALSE
)
}
\arguments{
\item{new_id}{id for loading the relevant metacell objects for scdb_mc and scdb_mat}

\item{smoothed_res}{Result object from smooth_zman_trajectory()}

\item{out_len}{The total of outcome observations along time, ie we usually set a number close to the total number of metacells to prevent over/under fitting}

\item{loess_degree}{Loess parameter to set the degree of fitting}

\item{loess_span}{Loess parameter to set spanning parameter}

\item{reverse_traj}{If the reverse coloring is preferred, set TRUE. Defaults to FALSE}
}
\description{
This function utilizes the smoothed trajectory to generate predicted expression along time
}
\section{Trajectory Analysis}{
NA
}

\examples{
filtered_GO_mc <- filter_exp(select_GO_mc, dispersion_threshold=0.05, threadnum = 1)
NK_smoothed_res = smooth_zman_trajectory(filtered_GO_mc, NK_mc_exprs$select_df, ref_k = 4)
NK_predicted_res = predict_expression_along_time(new_id, NK_smoothed_res, out_len = 36, loess_degree=1, loess_span = .9)
}
