% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zman_trajectory_analysis.R
\name{smooth_zman_trajectory}
\alias{smooth_zman_trajectory}
\title{Smooth Zman Trajectory}
\usage{
smooth_zman_trajectory(disp_go_df, normed_mc_res, ref_k = 4)
}
\arguments{
\item{disp_go_df}{Gene expression matrix after selection of GO genes and dispersion filter}

\item{normed_mc_res}{Result list from normalize_mc_exprs containing the raw AUC dataframe and corresponding metacell cell type labels}

\item{ref_k}{The number of k to build the reference path corresponding to the number of clusters, defaults to 4}
}
\description{
This function smoothes the AUC derived time of metacells to construct the Zman trajectory of the metacells
}
\section{Trajectory Analysis}{
NA
}

\examples{
# Smooth AUC trajectory:
# First we filter genes by dispersion to reduce the number of genes to more significant ones:
select_GO_mc <- get_GO_exp(select_exprs,gene_type = "gene_names", organism = "mouse", takelog=F)
# The threshold is set roughly at 90\% quantile of the mc, since metacells have fewer cells and pooled expression, we expect in general lower dispersion values
filtered_GO_mc <- filter_exp(select_GO_mc, dispersion_threshold=0.05, threadnum = 1)
NK_smoothed_res = smooth_zman_trajectory(filtered_GO_mc, NK_mc_exprs, ref_k = 4)
}
