% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{well_fcs_mc_Blood}
\alias{well_fcs_mc_Blood}
\title{Example data for timebin assignment}
\format{
An object of dataframe of metadata for cells
}
\usage{
well_fcs_mc_Blood
}
\description{
Example data for timebin assignment in blood
}
\keyword{datasets}
